+++ app/src/main/java/com/example/racestats/DigitalDash.kt
@@ -1,0 +1,4 @@
+package com.example.racestats
+
+class DigitalDash {
+}
