+++ app/src/main/java/com/example/racestats/EcuData.kt
@@ -1,0 +1,4 @@
+package com.example.racestats
+
+class EcuData {
+}
