+++ app/src/main/java/com/example/racestats/GMeter.java
@@ -1,0 +1,4 @@
+package com.example.racestats;
+
+public class GMeter {
+}
